move components into their own dir + some refactors
